chore: group gas fee updates file
